ENH: aggregate 활용 디비 쿼리문 작성

+++ server/graphql/types/index.js
@@ -62,7 +62,7 @@ export default `
   type Tour {
     name: String!
     price: Int!
-    rank: Float!
+    rank: Int!
     image: String!
     _id: String!
   }
